Moving utility methods into utils file.
